rolex: power: Support legacy powerHAL methods
Some interfaces are unable to directly communicate with the binderized
power HAL, either due to compiler limitations, or proprietary
limitations.

Our camera HAL for one, is unable to utilize the required c++14 & gnu++1z
syntax in order to communicate with the binderized interface as GCC 4.9
does not support them.

To workaround this, add a legacy module that translates legacy methods
and directs them to the new ones. In other words we're doing the
opposite of android.hardware power@1.0-impl; ironic isn't it?

Change-Id: I42192a52c57a4d4c9c052f3eebf9136837d1eb23